machine_simulator/requirements.txt
blinker
click
Faker
Flask
itsdangerous
Jinja2
MarkupSafe
paho-mqtt
python-dateutil
six
Werkzeug
certifi
